Cambios varios

+++ src/main/java/es/ucm/fdi/iw/g06/printopolis/model/SalesLine.java
@@ -36,7 +36,7 @@ public class SalesLine {
     private long design;
     private long sale;
     private int quantity;
-    private BigDecimal price;
+    private float price;
 
     private LocalDateTime date;
